Move action cache choose_provider to remote_provider
