+++ src/rust/engine/process_execution/remote/Cargo.toml
@@ -38,8 +38,6 @@ strum_macros = { workspace = true }
 parking_lot = { workspace = true }
 opendal = { workspace = true }
 remote_provider = { path = "../../remote_provider" }
-remote_provider_traits = { path = "../../remote_provider/remote_provider_traits" }
-remote_provider_opendal = { path = "../../remote_provider/remote_provider_opendal" }
 remote_provider_reapi = { path = "../../remote_provider/remote_provider_reapi" }
 
 [dev-dependencies]
