+++ src/rust/engine/remote_provider/Cargo.toml
@@ -6,6 +6,7 @@ authors = ["Pants Build <pantsbuild@gmail.com>"]
 publish = false
 
 [dependencies]
+grpc_util = { path = "../grpc_util" }
 remote_provider_opendal = { path = "./remote_provider_opendal" }
 remote_provider_reapi = { path = "./remote_provider_reapi" }
 remote_provider_traits = { path = "./remote_provider_traits" }
