+++ src/rust/engine/remote_provider/src/lib.rs
@@ -34,7 +34,7 @@ pub use remote_provider_traits::{
   RemoteOptions,
 };
 
-pub const REAPI_ADDRESS_SCHEMAS: [&str; 4] = ["grpc://", "grpcs://", "http://", "https://"];
+const REAPI_ADDRESS_SCHEMAS: [&str; 4] = ["grpc://", "grpcs://", "http://", "https://"];
 
 pub async fn choose_byte_store_provider(
   options: RemoteOptions,
@@ -67,6 +67,61 @@ pub async fn choose_byte_store_provider(
   } else {
     Err(format!(
       "Cannot initialise remote byte store provider with address {address}, as the scheme is not supported",
+    ))
+  }
+}
+
+pub async fn choose_action_cache_provider(
+  options: RemoteCacheProviderOptions,
+) -> Result<Arc<dyn ActionCacheProvider>, String> {
+  let address = options.action_cache_address.clone();
+
+  // TODO: we shouldn't need to gin up a whole copy of this struct; it'd be better to have the two
+  // set of remoting options managed together.
+  let remote_options = RemoteOptions {
+    cas_address: address.clone(),
+    instance_name: options.instance_name.clone(),
+    headers: options.headers.clone(),
+    tls_config: grpc_util::tls::Config::new(
+      options.root_ca_certs.clone(),
+      options.mtls_data.clone(),
+    )?,
+    rpc_timeout: options.rpc_timeout,
+    rpc_concurrency_limit: options.concurrency_limit,
+    // TODO: these should either be passed through or not synthesized here
+    chunk_size_bytes: 0,
+    rpc_retries: 0,
+    capabilities_cell_opt: None,
+    batch_api_size_limit: 0,
+  };
+
+  if REAPI_ADDRESS_SCHEMAS.iter().any(|s| address.starts_with(s)) {
+    Ok(Arc::new(
+      remote_provider_reapi::action_cache::Provider::new(options).await?,
+    ))
+  } else if let Some(path) = address.strip_prefix("file://") {
+    // It's a bit weird to support local "file://" for a 'remote' store... but this is handy for
+    // testing.
+    Ok(Arc::new(remote_provider_opendal::Provider::fs(
+      path,
+      "action-cache".to_owned(),
+      remote_options,
+    )?))
+  } else if let Some(url) = address.strip_prefix("github-actions-cache+") {
+    // This is relying on python validating that it was set as `github-actions-cache+https://...` so
+    // incorrect values could easily slip through here and cause downstream confusion. We're
+    // intending to change the approach (https://github.com/pantsbuild/pants/issues/19902) so this
+    // is tolerable for now.
+    Ok(Arc::new(
+      remote_provider_opendal::Provider::github_actions_cache(
+        url,
+        "action-cache".to_owned(),
+        remote_options,
+      )?,
+    ))
+  } else {
+    Err(format!(
+      "Cannot initialise remote action cache provider with address {address}, as the scheme is not supported",
     ))
   }
 }
